oled_controller/requirements.txt
cbor2==5.6.5
cffi==1.17.1
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
luma.core==2.4.2
luma.oled==3.14.0
markdown-it-py==3.0.0
mdurl==0.1.2
obs-cli==0.8.2
obs-websocket-py==1.0
obsws-python==1.7.1
pillow==11.2.1
psutil==7.0.0
pycparser==2.22
pyftdi==0.56.0
Pygments==2.19.1
pyserial==3.5
pyusb==1.3.1
rich==14.0.0
rpi-lgpio==0.6
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.7
websocket-client==1.8.0
dotenv==0.9.9
python-dotenv==1.1.0
